fix(rust): app's window size on monitors with a scale factor > 1
